refactor: move syntax type to a new file
